backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
redis==5.0.3
python-dotenv==1.0.1
pydantic-settings==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
pandas==2.2.1
numpy==1.26.4
alembic==1.13.1
